adding functions

+++ views/assets/script/userUiscript.js
@@ -113,29 +113,43 @@ $(document).on("click", "#add", () => {
 });
 
 $(document).on('click', '.delete', event => {
-    let id = $(event.target).data('id')
-    fetch(`http://${ip}:3000/users/${id}`, {
-        method: "DELETE",
-        headers: {
-            "Content-Type": "application/json",
-            "Access-Control-Allow-Origin": "*",
-            "Access-Control-Allow-Headers":
-                "Origin, X-Requested-With, Content-Type, Accept, Z-Key",
-            "Access-Control-Allow-Methods":
-                "GET, HEAD, POST, PUT, DELETE, OPTIONS",
-        },
-    }).then((response) => {
-        return response.text();
-    }).then((data) => {
-        Swal.fire({
-            title: "User deleted successfully!",
-            text: data,
-            icon: "warning",
-            confirmButtonText: "ok",
-        })
-    }).then(() => {
-        getDataUser.ajax.reload(null, false)
-    });
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: 'black',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            let id = $(event.target).data('id')
+            fetch(`http://${ip}:3000/users/${id}`, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    "Access-Control-Allow-Origin": "*",
+                    "Access-Control-Allow-Headers":
+                        "Origin, X-Requested-With, Content-Type, Accept, Z-Key",
+                    "Access-Control-Allow-Methods":
+                        "GET, HEAD, POST, PUT, DELETE, OPTIONS",
+                },
+            }).then((response) => {
+                return response.text();
+            }).then((data) => {
+                Swal.fire({
+                    title: "User delete !",
+                    text: data,
+                    icon: "warning",
+                    confirmButtonText: "ok",
+                })
+            }).then(() => {
+                getDataUser.ajax.reload(null, false)
+            });
+        } else if (result.isDismissed) {
+            console.log("cancelled");
+        }
+    })
 });
 
 $(document).on('click', '.edit', event => {
